clinical-classification/requirements.txt
# requirements.txt
# Clinical Note Classification System
# Python 3.11.0+

# ============================================
# Core Data Science & ML
# ============================================
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0

# ============================================
# OpenAI & LLM
# ============================================
openai>=1.12.0,<2.0.0
tiktoken>=0.5.0,<1.0.0

# ============================================
# Vector Database (Astra DB)
# ============================================
astrapy>=1.0.0,<2.0.0
langchain>=0.1.0,<0.3.0
langchain-astradb>=0.3.0,<1.0.0
langchain-openai>=0.0.8,<1.0.0
cassio>=0.1.4,<1.0.0

# ============================================
# Weak Supervision
# ============================================
snorkel>=0.9.9,<1.0.0

# ============================================
# Visualization
# ============================================
plotly>=5.18.0,<6.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
kaleido>=0.2.0,<1.0.0

# ============================================
# Dimensionality Reduction
# ============================================
umap-learn>=0.5.4,<1.0.0

# ============================================
# Progress Bars
# ============================================
tqdm>=4.66.0,<5.0.0

# ============================================
# Configuration & Environment
# ============================================
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0

# ============================================
# Utilities
# ============================================
requests>=2.31.0,<3.0.0
urllib3>=2.0.0,<3.0.0

# ============================================
# Testing (Optional)
# ============================================
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0

# ============================================
# Jupyter (Optional - for notebooks)
# ============================================
jupyter>=1.0.0,<2.0.0
ipykernel>=6.25.0,<7.0.0
ipywidgets>=8.1.0,<9.0.0

# ============================================
# Code Quality (Optional)
# ============================================
black>=23.7.0,<25.0.0
flake8>=6.1.0,<8.0.0
mypy>=1.5.0,<2.0.0